Move classes under new LeaseManagement folder/namespace, for both shipping and test code. No functional changes at all, source code just gets namespace name changes.
