+++ src/DocumentDB.ChangeFeedProcessor.UnitTests/ChangeFeedProcessorTests.cs
@@ -9,15 +9,14 @@ namespace Microsoft.Azure.Documents.ChangeFeedProcessor.UnitTests
     using System.Linq;
     using System.Threading;
     using System.Threading.Tasks;
-    using Microsoft.Azure.Documents.ChangeFeedProcessor.Bootstrapping;
     using Microsoft.Azure.Documents.ChangeFeedProcessor.DataAccess;
     using Microsoft.Azure.Documents.ChangeFeedProcessor.FeedProcessing;
+    using Microsoft.Azure.Documents.ChangeFeedProcessor.LeaseManagement;
     using Microsoft.Azure.Documents.ChangeFeedProcessor.PartitionManagement;
     using Microsoft.Azure.Documents.ChangeFeedProcessor.UnitTests.Utils;
     using Microsoft.Azure.Documents.Client;
     using Microsoft.Azure.Documents.Linq;
     using Moq;
-    using Newtonsoft.Json;
     using Xunit;
 
     [Trait("Category", "Gated")]
