+++ src/DocumentDB.ChangeFeedProcessor.UnitTests/Exceptions/LeaseLostExceptionTests.cs
@@ -1,4 +1,9 @@
-﻿using Microsoft.Azure.Documents.ChangeFeedProcessor.Exceptions;
+﻿//----------------------------------------------------------------
+// Copyright (c) Microsoft Corporation.  Licensed under the MIT license.
+//----------------------------------------------------------------
+
+using Microsoft.Azure.Documents.ChangeFeedProcessor.Exceptions;
+using Microsoft.Azure.Documents.ChangeFeedProcessor.LeaseManagement;
 using Microsoft.Azure.Documents.ChangeFeedProcessor.PartitionManagement;
 using Moq;
 using System;
