+++ src/DocumentDB.ChangeFeedProcessor.UnitTests/FeedProcessor/PartitionProcessorFactoryTests.cs
@@ -4,6 +4,7 @@
     using AutoFixture;
     using Microsoft.Azure.Documents.ChangeFeedProcessor.DataAccess;
     using Microsoft.Azure.Documents.ChangeFeedProcessor.FeedProcessing;
+    using Microsoft.Azure.Documents.ChangeFeedProcessor.LeaseManagement;
     using Microsoft.Azure.Documents.ChangeFeedProcessor.PartitionManagement;
     using Microsoft.Azure.Documents.Client;
     using Moq;
