+++ src/DocumentDB.ChangeFeedProcessor/DocumentDB.ChangeFeedProcessor.csproj
@@ -70,7 +70,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Folder Include="LeaseManagement\" />
     <Folder Include="Properties\" />
   </ItemGroup>
 </Project>
