+++ src/DocumentDB.ChangeFeedProcessor/LeaseManagement/DocumentServiceLease.cs
@@ -2,12 +2,13 @@
 // Copyright (c) Microsoft Corporation.  Licensed under the MIT license.
 //----------------------------------------------------------------
 
-namespace Microsoft.Azure.Documents.ChangeFeedProcessor.PartitionManagement
+namespace Microsoft.Azure.Documents.ChangeFeedProcessor.LeaseManagement
 {
     using System;
     using System.Collections.Generic;
     using System.Globalization;
     using Microsoft.Azure.Documents;
+    using Microsoft.Azure.Documents.ChangeFeedProcessor.PartitionManagement;
     using Newtonsoft.Json;
 
     [Serializable]
