+++ src/DocumentDB.ChangeFeedProcessor/LeaseManagement/FixedCollectionLeaseManagerFactory.cs
@@ -4,7 +4,6 @@
 
 namespace Microsoft.Azure.Documents.ChangeFeedProcessor.LeaseManagement
 {
-    using Microsoft.Azure.Documents.ChangeFeedProcessor.DataAccess;
     using Microsoft.Azure.Documents.ChangeFeedProcessor.PartitionManagement;
     using Microsoft.Azure.Documents.Client;
