+++ src/DocumentDB.ChangeFeedProcessor/LeaseManagement/ILeaseManagerFactory.cs
@@ -4,8 +4,6 @@
 
 namespace Microsoft.Azure.Documents.ChangeFeedProcessor.LeaseManagement
 {
-    using Microsoft.Azure.Documents.ChangeFeedProcessor.PartitionManagement;
-
     internal interface ILeaseManagerFactory
     {
         ILeaseManager Create(LeaseManagerParameters parameters);
