+++ src/DocumentDB.ChangeFeedProcessor/PartitionManagement/PartitionLoadBalancer.cs
@@ -8,7 +8,7 @@ namespace Microsoft.Azure.Documents.ChangeFeedProcessor.PartitionManagement
     using System.Collections.Generic;
     using System.Threading;
     using System.Threading.Tasks;
-
+    using Microsoft.Azure.Documents.ChangeFeedProcessor.LeaseManagement;
     using Microsoft.Azure.Documents.ChangeFeedProcessor.Logging;
 
     internal class PartitionLoadBalancer : IPartitionLoadBalancer
