+++ src/DocumentDB.ChangeFeedProcessor/PartitionManagement/PartitionSupervisorFactory.cs
@@ -6,6 +6,7 @@ namespace Microsoft.Azure.Documents.ChangeFeedProcessor.PartitionManagement
 {
     using System;
     using Microsoft.Azure.Documents.ChangeFeedProcessor.FeedProcessing;
+    using Microsoft.Azure.Documents.ChangeFeedProcessor.LeaseManagement;
 
     internal class PartitionSupervisorFactory : IPartitionSupervisorFactory
     {
